pyproject.toml
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "autopep8>=2.0.1",
]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Mikołaj Nowak", email = "miknowak@outlook.com"},
]
dependencies = [
    "pyautogui>=0.9.53",
    "python-gnupg>=0.5.0",
    "pydantic>=1.10.5",
]
requires-python = ">=3.11"
license = {text = "All rights reserved"}
